feature: schedule service complete — CRUD, export/import PDF/XLSX, dispatcher role, nginx DNS fix, Toaster
- ScheduleService: Full CRUD with overlap validation (UC-15)
- ScheduleExport: PDF (QuestPDF) + XLSX (ClosedXML) (UC-17)
- ScheduleImport: XLSX upload with validation + conflict detection (UC-18)
- ScheduleEntryDialog + ScheduleImportDialog + AlertDialog
- ScheduleFilterBar + ScheduleTable with export/import/add buttons
- Dispatcher role: seed data, quick-login redirect to /schedule
- Toaster (sonner) added to root layout — toasts now visible
- extractErrorMessage helper — handles 3 error formats (Result<T>, ProblemDetails, ErrorResponse)
- nginx: resolver directive + set variables — prevents 502 on container restart
- Dockerfile: NuGet cache mount for fast rebuilds
- Teacher/Group GET endpoints: all roles allowed
- ScheduleResponse.DayOfWeek: enum → int for JSON compatibility
- User stories updated: all 7 ScheduleService UC marked ✅

+++ CollegeLMS.API/Mappers/ScheduleMapper.cs
@@ -16,7 +16,7 @@ public static class ScheduleMapper
             TeacherName = entry.Teacher?.User?.FullName,
             Subject = entry.Subject,
             Room = entry.Room,
-            DayOfWeek = entry.DayOfWeek,
+            DayOfWeek = (int)entry.DayOfWeek,
             StartTime = entry.StartTime,
             EndTime = entry.EndTime,
             LessonType = entry.LessonType.ToString(),
